requirements.txt
aiosqlite==0.18.0
anyio==3.6.2
backports.zoneinfo==0.2.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
fastapi-users==10.4.1
fastapi-users-db-sqlalchemy==5.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
httpx-oauth==0.11.0
idna==3.4
itsdangerous==2.1.2
makefun==1.15.1
passlib==1.7.4
psycopg==3.1.8
psycopg-binary==3.1.8
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
